Add -Wall -Wextra for clang and gcc and clean up resulting errors. Note that this doesn't enable -Weverything for clang. I don't have an objection to doing so, but I think we should discuss before turning that on since that's going to result in potential build faiures for running a newer version of clang than us. Maybe we could have a build option with -Weverything as part of our TBD lint process.

+++ src/CodeGen/JumpTable.cpp
@@ -94,7 +94,7 @@ namespace NativeJIT
     // WARNING: Non portable. Assumes that fixup value is labelAddress - siteAddress - size.Size().
     void JumpTable::PatchCallSites()
     {
-        for (int i=0; i < m_callSites.size(); ++i)
+        for (size_t i=0; i < m_callSites.size(); ++i)
         {
             const CallSite& site = m_callSites[i];
             const uint8_t* labelAddress = AddressOfLabel(site.GetLabel());
